+++ apps/src/sites/studio/pages/certificates/show.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import getScriptData from '@cdo/apps/util/getScriptData';
-import CertificateShare from '@cdo/apps/templates/CertificateShare';
+import CertificateShare from '@cdo/apps/templates/certificates/CertificateShare';
 import {getStore} from '@cdo/apps/redux';
 
 $(document).ready(function() {
